Greatly improved README.md and made some improvements to build.sh

+++ build.sh
@@ -1,43 +1,58 @@
-SEPERATOR="---------------------------------------------------";
-echo "[ABOUT] Security System program by Christian Lisle (http://christianlisle.com)"
-echo "[ABOUT] Learn More about this project: https://github.com/ChristianLisle/Security-System"
+red=`tput setaf 1`
+green=`tput setaf 2`
+reset=`tput sgr0`
 echo "[ABOUT] This bash script constructs necessary storage arrangements and ensures that the local environment is suitable to run the program."
-# Build the environment folders and files needed for storing data
-echo "${SEPERATOR}\n[BUILD] Constructing data storage directories, setting data values, and creating essential storage files."
 
-directories=("data" "data/recordings" "data/stream_frames")
-
-for directory in ${directories[@]}
+# Ask the user if they are building the environment for the client or server side. Client side does not need all of the same packages or storage directories as server side.
+echo "[PROMPT] Are you building the environment for the client or server side?"
+while true;
 do
-  DIR="${directory}"
-  if [ ! -d "$DIR" ]; then
-    echo "[BUILD] Creating directory: ${directory}"
-    mkdir "${directory}"
+  read -p "[PROMPT] Enter \"client\" or \"server\": " choice
+  if [ $choice == "client" ] || [ $choice == "Client" ]; then
+    deploy="CLIENT"
+    break
+  elif [ $choice == "server" ] || [ $choice == "Server" ]; then
+    deploy="SERVER"
+    # Build the environment folders and files needed for storing data (Server side only).
+    echo "[${deploy} BUILD] Constructing data storage directories, setting data values, and creating essential storage files."
+
+    directories=("data" "data/recordings" "data/stream_frames")
+
+    for directory in ${directories[@]}
+    do
+      DIR="${directory}"
+      if [ ! -d "$DIR" ]; then
+        echo "[${deploy} BUILD] Creating directory: ${directory}"
+        mkdir "${directory}"
+      fi
+    done
+
+    ENV_FILE=.env
+    if test -f "$ENV_FILE"; then
+      max=$(grep MAX_CLIENTS .env | cut -d '=' -f2)
+    else
+      max=5
+      echo "[${deploy} BUILD] .env file not found.\n[${deploy} BUILD] Construcing new .env file with a default MAX_CLIENTS of 5."
+      echo "MAX_CLIENTS=${max}\nGMAIL_USER=\nGMAIL_APP_PASSWORD=" > .env
+    fi
+
+    for i in `seq 1 $max`
+    do
+      if [ ! -d "${directories[2]}/${i}" ]; then
+        echo "[${deploy} BUILD] Creating storage directory: data/stream_frames/${i}"
+        mkdir data/stream_frames/$i
+      fi
+    done
+
+    touch data/blacklist.txt
+    echo "on" > data/alarm_status.txt
+    echo "0" > data/clients.txt
+    break
   fi
 done
 
-ENV_FILE=.env
-if test -f "$ENV_FILE"; then
-  max=$(grep MAX_CLIENTS .env | cut -d '=' -f2)
-else
-  max=5
-  echo "[BUILD] .env file not found.\n[BUILD] Construcing new .env file with a default MAX_CLIENTS of 5."
-  echo "MAX_CLIENTS=${max}\nGMAIL_USER=\nGMAIL_APP_PASSWORD=" > .env
-fi
 
-for i in `seq 1 $max`
-do
-  if [ ! -d "${directories[2]}/${i}" ]; then
-    echo "[BUILD] Creating storage directory: data/stream_frames/${i}"
-    mkdir data/stream_frames/$i
-  fi
-done
-
-touch data/blacklist.txt
-echo "on" > data/alarm_status.txt
-echo "0" > data/clients.txt
-
-echo "${SEPERATOR}\n[BUILD] Checking Python version"
+echo "[${deploy} BUILD] Checking Python version"
 pvs="$(python3 --version)"
 IFS='.' read -r -a version <<< "${pvs:7:20}" # convert Python version string to array (i.e., "3.8.5" to 3 8 5)
 
@@ -54,30 +69,49 @@ elif [[ $((${version[0]} + 0)) -gt 3 ]]; then
 fi
 
 if [ valid ]; then
-  echo "[BUILD] Your Python version (${pvs:7:10}) is sufficient."
+  echo "[${deploy} BUILD] Your Python version (${pvs:7:10}) is sufficient."
 else
-  echo "[BUILD] Your Python version (${pvs:7:10}) is older than the recommended version, Python 3.8.5. Consider updating before running the program."
+  echo "[${deploy} BUILD] ${red}WARNING${reset}: Your Python version (${pvs:7:10}) is older than the recommended version, Python 3.8.5. Consider updating before running the program."
 fi
 
 # Install python module dependencies.
-echo "${SEPERATOR}\n[BUILD] Installing Python modules using pip. ${WARNING}"
+echo "[${deploy} BUILD] Installing Python modules using pip. ${WARNING}"
 
 # Install pip commands using pip3 or pip. If neither commands work, alert the user.
 {
   pip3 install --upgrade pip
   pip3 install opencv-python-headless
   pip3 install python-decouple
-  pip3 install flask
-  pip3 install waitress
+  if [ $deploy == "SERVER" ]; then
+    pip3 install flask
+    pip3 install waitress
+  fi
 } || {
   pip install --upgrade pip
   pip install opencv-python-headless
   pip install python-decouple
-  pip install flask
-  pip install waitress
-}|| {
-  echo "[FAILURE] Could not install Python modules using pip. Manually install Python modules listed on the README. Your environment should then be ready."
+  if [ $deploy == "SERVER" ]; then
+    pip install flask
+    pip install waitress
+  fi
+} || {
+  python -m pip install --upgrade pip
+  python -m pip install opencv-python-headless
+  python -m pip install python-decouple
+  if [ $deploy == "SERVER" ]; then
+    python -m pip install flask
+    python -m pip install waitress
+  fi
+} || {
+  python3 -m pip install --upgrade pip
+  python3 -m pip install opencv-python-headless
+  python3 -m pip install python-decouple
+  if [ $deploy == "SERVER" ]; then
+    python3 -m pip install flask
+    python3 -m pip install waitress
+  fi
+} || {
+  echo "${red}[BUILD FAILURE]${reset} Could not install Python modules using pip. View step 3 on README.md for more information on installing Pip or view step 4 on README.md to manually install the required Python packages."
   exit
 }
-
-echo "[SUCCESS] Your environment appears to be sufficient for this program. View the README (https://github.com/ChristianLisle/Security-System) for information on how to run the program."
+echo "${green}[BUILD SUCCESS]${reset} Your environment appears to be sufficient for this program. View step 4 on README.md for information on how to run the program."
